backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "darkroom-backend"
version = "0.1.0"
description = "Darkroom — local open-source video editor backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "openai-whisper>=20231117",
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",  # async test client for FastAPI
]

[project.scripts]
darkroom = "darkroom.main:run"

[tool.hatch.build.targets.wheel]
packages = ["darkroom"]
